feat: Add `special_account` to `Account` and options
This is to help address "semi-firm" Account/data problem. Think this
might be a good solution. Hope to list more `SPECIAL` accounts in
future, plenty of other candidates -- anything that needs calculating.

Not completely happy with `payments` as own field, but the need for this
subset of the accounts/data is very pervasive, can't help but think will
be long-run helpful, it just looks ugly.