refactor: Rearrange Entry create_object methods and tests
As expected from 180f2992932c89b4fe1b472c4248a7acb7010cc2, rearranging
of Entry methods to make more sense. Not there yet but closer.

In this case try and put in own methods actions for any particular line.

Broke the most challenging tests, but this is not a bad thing, those
tests were always too persnickety and a bit of a waste of time. End-to-
end is all that matters.