major fix: Attend to problem of superfluous Transaction creation
To ensure `Transaction` objects are not created if there is a validation
problem with the `Entry`, as `Transaction` must be created before
`Entry`.

This is important for system integrity, until there is some generic
relation from `Transaction` that we can ensure exists.

@@ TODO: This important integrity check needs to be donebetter. This
should nearly certainly be done as post_save signal on Entry. taiga#122

In this commit doing some things to make tighter and to setup post_save
signal:
- adding `Integrity` constraint on `Transaction.source`
- removing `Integrity` constraint on `Entry.transaction`
   (so entry can be created before transaction)
- disallowing create Transaction in admin

NB: reinitialise migrations again. Hopefully this is the last time.