Refactor `BankTransaction` to `BankLine`
`ledgers.Transaction` is one of the most important object classes in the
application.

`BankTransaction` is one of the only classes so far that does NOT refer
to `ledgers` in some way. It is just a dump of the statement, a
container for external information. Became apparent while revising
comments that having `Transaction` in the name is conflating concepts.

This naming is inviting confusion.

Renamed now before it gets harder to do later.